# lightsim2grid: changed checks for lightsim2grid