--- a/pandapower/auxiliary.py
+++ b/pandapower/auxiliary.py
@@ -736,8 +736,7 @@ def _check_if_numba_is_installed(numba):
     return numba
 
 
-def _check_lightsim2grid_compatibility(net, lightsim2grid, voltage_dependend_loads, algorithm, enforce_q_lims,
-                                       distributed_slack):
+def _check_lightsim2grid_compatibility(net, lightsim2grid, voltage_depend_loads, algorithm, distributed_slack):
     """
     Implement some checks to decide whether the package lightsim2grid can be used. The package implements a backend for
      power flow calculation in C++ and provides a speed-up. If lightsim2grid is "auto" (default), we don't bombard the
@@ -755,35 +754,17 @@ def _check_lightsim2grid_compatibility(net, lightsim2grid, voltage_dependend_loa
         if lightsim2grid == "auto":
             return False
         raise NotImplementedError(f"option 'lightsim2grid' is True but the algorithm {algorithm} not implemented.")
-    if voltage_dependend_loads:
-        if lightsim2grid != "auto":
-            logger.info("option 'lightsim2grid' is True and voltage-dependent loads detected. Falling back to "
-                        "pandapower implementation")
-        return False
-    #if enforce_q_lims:
-    # not valid: the way you enforce the q_limits in pandapower is to loop and perform successive calls to the
-    # newtonpf function, in this case using the lightsim2grid equivalent is perfectly fine. You CAN use
-    # lightsim2grid newtonpf function in this case.
-    #    logger.info("option 'lightsim2grid' is True and enforce_q_lims is True. Falling back to pandapower "
-    #                "implementation")
-    #    return False
+    if voltage_depend_loads:
+        if lightsim2grid == "auto":
+            return False
+        raise NotImplementedError("option 'lightsim2grid' is True and voltage-dependent loads detected.")
     if len(net.ext_grid) > 1 and not distributed_slack:
         # lightsim2grid implements distributed_slack similarly to pandapower, but does not accept multiple slacks
         # otherwise
-        if lightsim2grid != "auto":
-            logger.info("option 'lightsim2grid' is True and multiple ext_grids are found. Falling back to "
-                        "pandapower implementation.")
-        return False
-    # if np.any(net.gen.bus.isin(net.ext_grid.bus)):
-    #     #  not valid: actually it's even the exact opposite. In lightsim2grid, only the generators are considered
-    #     #  as slack buses. There is no concept of "ext_grid" in lightsim2grid. And by the way, this is not
-    #     #  relevant
-    #     #  when using the newtonpf function, ext_grid and generators with slack bus are handled before that when
-    #     #  defining the pv , pq and ref index vectors (used in newtonpf). You CAN use lightsim2grid newtonpf
-    #     #  function in this case.
-    #     logger.info("option 'lightsim2grid' is True and gen elements were found at slack buses. Falling back "
-    #                 "to pandapower implementation")
-    #     return False
+        if lightsim2grid == "auto":
+            return False
+        raise NotImplementedError("option 'lightsim2grid' is True and multiple ext_grids are found, "
+                                  "but distributed_slack=False.")
 
     return True
 
@@ -1006,7 +987,7 @@ def _init_runpp_options(net, algorithm, calculate_voltage_angles, init,
             voltage_depend_loads = False
 
     lightsim2grid = _check_lightsim2grid_compatibility(net, lightsim2grid, voltage_depend_loads, algorithm,
-                                                       enforce_q_lims, distributed_slack)
+                                                       distributed_slack)
 
     ac = True
     mode = "pf"