--- a/pandapower/pf/makeYbus_numba.py
+++ b/pandapower/pf/makeYbus_numba.py
@@ -166,4 +166,7 @@ def makeYbus(baseMVA, bus, branch):
     Yx, Yj, Yp, nnz = gen_Ybus(Yf_x, Yt_x, Ysh, col_Y, f, t, np.argsort(f), np.argsort(t), nb, nl,
                                np.arange(nl, dtype=np.int64))
     Ybus = csr_matrix((np.resize(Yx, nnz), np.resize(Yj, nnz), Yp), (nb, nb))
+    for Y in (Ybus, Yf, Yt):
+        Y.sort_indices()
+        Y.eliminate_zeros()
     return Ybus, Yf, Yt