--- a/pandapower/pf/runpp_3ph.py
+++ b/pandapower/pf/runpp_3ph.py
@@ -407,7 +407,7 @@ def runpp_3ph(net, calculate_voltage_angles=True, init="auto",
     _add_pf_options(net, tolerance_mva=tolerance_mva, trafo_loading=trafo_loading,
                     numba=numba, ac=ac, algorithm="nr", max_iteration=max_iteration,\
                     only_v_results=only_v_results,v_debug=v_debug, use_umfpack=use_umfpack,
-                    permc_spec=permc_spec)
+                    permc_spec=permc_spec, lightsim2grid=False)
     net._options.update(overrule_options)
     _check_bus_index_and_print_warning_if_high(net)
     _check_gen_index_and_print_warning_if_high(net)