@BDonnot 2022-01-24
Hello,

Can I have a way to import the "*example_simple()*"  function ? I cannot reproduce the bug here without this function :-/