@rbolgaryn 2022-01-24
Hi @BDonnot ,

the example_simple function is in pandapower.networks (should be imported extra):

´´´python
import pandapower as pp
import pandapower.networks
net = pp.networks.example_simple()
´´´

Roman