@BDonnot 2022-01-24
Hello,

I cannot reproduce the problem on my dev machine (ubuntu 20.04, python3.8)

Can you launch:

```python
print("now with example provided")
import pandapower as pp
import lightsim2grid
import pdb
import pandas as pd
from pandapower.networks import example_simple
import numpy as np
import pandas as pd
print(f"{np.__version__ = }")
print(f"{pd.__version__ = }")
print(f"{lightsim2grid.__version__ = }")
print(f"{pp.__version__ = }")
net = example_simple()
pp.runpp(net, distributed_slack=True)  # works
pp.runpp(net, distributed_slack=True, numba=False)  # LoadFlowNotConverged
assert net.converged

#second
pp.runpp(net, distributed_slack=True)  # works
pp.runpp(net, distributed_slack=True, numba=False, init="results")  # works
assert net.converged

#third
net2 = example_simple()
pp.runpp(net, distributed_slack=True)  # works
pp.runpp(net2, distributed_slack=True, lightsim2grid=False)  # works
assert net2.converged
# assert_res_equal(net, net2)  # works  # -> not imported so skipping it
assert net.res_gen.equals(net2.res_gen)
assert net.res_load.equals(net2.res_load)
assert net.res_line.equals(net2.res_line)
print("everything match !")
```

And give me the output ? 

Also, can you give me your setup (os and python version please)

FYI for me the above script gives:

```
now with example provided
np.__version__ = '1.21.5'
pd.__version__ = '1.4.0'
lightsim2grid.__version__ = '0.6.0'
pp.__version__ = '2.7.0'
/home/benjamin/Documents/pp_test/venv/lib/python3.8/site-packages/pandapower/create.py:509: FutureWarning: pandas.Int64Index is deprecated and will be removed from pandas in a future version. Use pandas.Index with the appropriate dtype instead.
  net[s] = pd.DataFrame(zeros(0, dtype=net[s]), index=pd.Int64Index([]))
...
[Above warning at least 50 times !]
everything match !
 ```

Also it's unclear for me what you want to test exactly.

Your are totally correct, lightsim2grid does not rely on numba at all. If the flag "numba=True" changes something, then it's probably something given as input of lightsim2grid, for example the init voltages.