@rbolgaryn 2022-01-24
I have this output:

```python
now with example provided
np.__version__ = '1.21.5'
pd.__version__ = '1.3.5'
lightsim2grid.__version__ = '0.6.0'
pp.__version__ = '2.7.1.dev1'
Traceback (most recent call last):
  File "C:\ProgramData\Miniconda3\envs\pandapower\lib\site-packages\IPython\core\interactiveshell.py", line 3251, in run_code
    exec(code_obj, self.user_global_ns, self.user_ns)
  File "<ipython-input-2-12c50577f436>", line 15, in <module>
    pp.runpp(net, distributed_slack=True, numba=False)  # LoadFlowNotConverged
  File "C:\Users\rbolgaryn\repos\pandapower\pandapower\run.py", line 231, in runpp
    _powerflow(net, **kwargs)
  File "C:\Users\rbolgaryn\repos\pandapower\pandapower\powerflow.py", line 86, in _powerflow
    _ppci_to_net(result, net)
  File "C:\Users\rbolgaryn\repos\pandapower\pandapower\powerflow.py", line 188, in _ppci_to_net
    raise LoadflowNotConverged("Power Flow {0} did not converge after "
pandapower.powerflow.LoadflowNotConverged: Power Flow nr did not converge after 10 iterations!
```