@rbolgaryn 2022-01-24
My Python version is 3.9 on Windows 10. 
Also, the tests fail on the GitHub Actions servers, which must be linux.

The difference in behavior is very confusing.