@BDonnot 2022-01-24
Oh ok it was my mistake, I did not installed your pandapower version by the default one and now i'm able to reproduce it and I get the error:

```
  File "bug_pp.py", line 12, in <module>
    pp.runpp(net, distributed_slack=True, numba=False)  # LoadFlowNotConverged
  File "pp_test/pandapower/run.py", line 233, in runpp
    _powerflow(net, **kwargs)
  File "pp_test/pandapower/powerflow.py", line 86, in _powerflow
    _ppci_to_net(result, net)
  File "pp_test/pandapower/powerflow.py", line 188, in _ppci_to_net
    raise LoadflowNotConverged("Power Flow {0} did not converge after "
pandapower.powerflow.LoadflowNotConverged: Power Flow nr did not converge after 10 iterations!
```

I'll try to see what is the cause of it