@rbolgaryn 2022-01-24
I also see the same behavior in linux, with Python 3.9.9:

now with example provided
np.__version__ = '1.21.5'
pd.__version__ = '1.4.0'
lightsim2grid.__version__ = '0.6.0'
pp.__version__ = '2.7.1.dev1'