@BDonnot 2022-01-24
As i thought, the difference is in pandapower and not in lightsim2grid.

The `Ybus` that is initialized (given as input in the newton functions) is not exactly the same if you pass `numba=True` or `numba=False`

When it is initialized with `numba=False` it appears the Ybus is not correct and has some issues.

For example, just by printing it:

```python
def _run_ac_pf_without_qlims_enforced(ppci, options):
    makeYbus, pfsoln = _get_numba_functions(ppci, options)

    baseMVA, bus, gen, branch, ref, pv, pq, _, _, V0, ref_gens = _get_pf_variables_from_ppci(ppci)

    ppci, Ybus, Yf, Yt = _get_Y_bus(ppci, options, makeYbus, baseMVA, bus, branch)

    # compute complex bus power injections [generation - load]
    Sbus = _get_Sbus(ppci, options["recycle"])


    # run the newton power flow
    import numpy as np
    print(f"{np.sum(np.abs(Ybus)) = }")  # the line I added

    [rest of the function that i do not copy !]
```

And this simple snippet works:

```python
import pandapower as pp
from pandapower.networks import example_simple
pp.runpp(net, distributed_slack=True, numba=False, lightsim2grid=True)  # works only with the "print" case
pp.runpp(net, distributed_slack=True, numba=True, lightsim2grid=True)  # works in both cases
```

I have absolutely no idea why printing something changes the behaviour...