@rbolgaryn 2022-01-25
I can confirm that the snippet also works for me