@BDonnot 2022-01-26
It's still super weird, but I guess keeping it this way won't hurt pandapower too much.

I don't have any clue on what this does under the hood to solve the issue...