@rbolgaryn 2022-01-26
There are still some failing tests, especially in the distributed slack functionality. I will take a look further