@rbolgaryn 2022-01-26
@BDonnot ,

what is the used sparce matrix encoding in lightsim2grid? (CSR or CRS?)