@BDonnot 2022-01-26
Information is here: https://lightsim2grid.readthedocs.io/en/latest/use_solver.html 

`Ybus = ...  # a csc sparse matrix (it's really important that it is a csc and not a csr !)`