@rbolgaryn 2022-01-27
I can try initializing it as a csc if lightsim2grid is used (for numba=False)