@rbolgaryn 2022-02-02
Here some timings:

```python
import pandapower as pp
import pandapower.networks

# case9241pegase:
net = pp.networks.case9241pegase()
pp.runpp(net, numba=True, lightsim2grid=False)

# without lightsim2grid
%timeit pp.runpp(net, numba=True, lightsim2grid=False)
340 ms ± 13.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

net._ppc["et"]
0.313649500021711


# with lightsim2grid
%timeit pp.runpp(net, numba=True, lightsim2grid=True)
127 ms ± 2.66 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
net._ppc["et"]
0.10265469993464649


# case39:
net = pp.networks.case39()

# without lightsim2grid:
%timeit pp.runpp(net, numba=True, lightsim2grid=False)
9.02 ms ± 60.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
net._ppc["et"]
0.0033928999910131097

# with lightsim2grid:
%timeit pp.runpp(net, numba=True, lightsim2grid=True)
8.26 ms ± 236 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
net._ppc["et"]
0.0023731000255793333
```

Seems that the benefit in the performance is there, but not in the same magnitude as shown in benchmarks of lightsim2grid.

The explanation is probably that the overhead in pandapower is still quite seizable, and it weighs on the total execution time (the performance boost for a small grid is not very high, but with a large grid is twice as fast - supports this point).
The "et" time (power flow time as returned by the newtonpf function) without lightsim2grid is ~3 times longer for the large grid, and not very much different for the small grid.