@BDonnot 2022-02-02
There are still huge penalty in the timings there.

For example if I launch this code:

```python
import pandapower as pp
import pandapower.networks
import time
import sys

nb_repeat = 10
# case9241pegase:
for nm_ in ["case39", "case118", "case9241pegase"]:
    net = getattr(pp.networks, nm_) () #.case9241pegase()

    # "compile" it with numba (first call always slower)
    pp.runpp(net, init="flat", numba=True, lightsim2grid=False)
    print(f"for {nm_}")
    beg_ = time.process_time()
    for _ in range(nb_repeat):
        pp.runpp(net, init="flat", numba=True, lightsim2grid=False)
    end_ = time.process_time()
    print(f"\t without ls: {1000. * (end_- beg_) / nb_repeat :.2f} ms")
    # without lightsim2grid

    beg_ = time.process_time()
    for _ in range(nb_repeat):
        pp.runpp(net, init="flat", numba=True, lightsim2grid=True)
    end_ = time.process_time()
    print(f"\t with ls: {1000. * (end_- beg_) / nb_repeat:.2f} ms")

```

And I modify the "newton_pf" function just print  the time spent in "solver.solve" like this:

```python
def newtonpf_new(Ybus, Sbus, V0, ref, pv, pq, ppci, options):
    max_iteration = options["max_iteration"]
    tolerance_pu = options['tolerance_mva']  # / ppci["baseMVA"]

    try:
        # lazy import for earlier pandapower version (without distributed slack):
        from pandapower.pypower.idx_bus import SL_FAC
        # contribution factors for distributed slack:
        slack_weights = ppci['bus'][:, SL_FAC].astype(np.float64)
    except ImportError:
        # earlier version of pandapower
        warnings.warn("You are using a pandapower version that does not support distributed slack. We will attempt to "
                      "replicate this with lightsim2grid.")
        ref, slack_weights = _isolate_slack_ids(Sbus, pv, pq)
    
    # initialize the solver and perform some sanity checks
    solver = _get_valid_solver(options, Ybus)
    
    # do the newton raphson algorithm
    import time
    beg_ = time.perf_counter()
    solver.solve(Ybus, V0, Sbus, ref, slack_weights, pv, pq, max_iteration, tolerance_pu)
    end_ = time.perf_counter()
    
    # extract the results
    Va = solver.get_Va()
    Vm = solver.get_Vm()
    V = Vm * np.exp(1j * Va)
    J = solver.get_J()
    converged = solver.converged()
    iterations = solver.get_nb_iter()
    print(f"\t\t{end_ - beg_}")

    Vm_it = None
    Va_it = None
    return V, converged, iterations, J, Vm_it, Va_it
```

I get the following results:
```
for case39
	 without ls: 11.77 ms
		0.0002456940710544586
		0.00021251291036605835
		0.00021051301155239344
		0.00022260495461523533
		0.00020686304196715355
		0.00022080691996961832
		0.00024157599546015263
		0.00020660599693655968
		0.0002070080954581499
		0.00021298404317349195
	 with ls: 10.38 ms
for case118
	 without ls: 14.05 ms
		0.00048077991232275963
		0.00047814392019063234
		0.00048135698307305574
		0.0004840519977733493
		0.00048554898239672184
		0.0004892740398645401
		0.0004753390094265342
		0.00047354702837765217
		0.00047925999388098717
		0.0004946120316162705
	 with ls: 11.81 ms
for case9241pegase
	 without ls: 390.22 ms
		0.06634897598996758
		0.06326793599873781
		0.0629391330294311
		0.06273389200214297
		0.06280133000109345
		0.06275211391039193
		0.06393996707629412
		0.06325664499308914
		0.0629483190132305
		0.06257649604231119
	 with ls: 113.96 ms
```

So basically, the time spent to perform the powerflow (for case39 for example) is around 0.2ms (*eg* 0.0002456940710544586 s)

On your measure it was 2ms (so 10x more) I'll try to see what causes the difference, but for sure there's something happening.